% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalat.R
\name{getPalat}
\alias{getPalat}
\title{Get the color palette}
\usage{
getPalat(df, hex = "HEX", labels = NULL, plot = TRUE, ...)
}
\arguments{
\item{df}{A data frame containing color information.}

\item{hex}{A string specifying the column name for hex color codes. Default "HEX". You also might use "HEX.K"}

\item{labels}{A string specifying the column name for labels. Default NULL.}

\item{plot}{Logical. Whether the palette is to be visualized. Default TRUE.}

\item{...}{Other arguments passed to geom_tile() and geom_text()}
}
\value{
A named character vector of the color codes and labels if provided.
}
\description{
Get the color palette
}
\examples{
df <- data.frame(region = c("Andes", "Caribe"), HEX = c("#00FFD7", "#9A649F"))
getPalat(df, labels = "region")

# Sample data frame
df <- data.frame(
region = c("Andes", "Andes", "Andes", "Caribe", "Caribe"),
city = c("Bogota", "Medellin", "Cali", "Barranquilla", "Santa Marta"),
latitude = c(4.611, 6.251, 3.437, 10.391, 10.963),
longitude = c(-74.083, -75.563, -76.522, -75.514, -74.796))

# Get the colors in the same df for each city
df <- getColors(df, lat = "latitude", lon = "longitude", map = FALSE)
palatXcity <- getPalat (df, labels = "city")
print(palatXcity)

# Get the colors in the same df for each region
df <- colorCluster(df, k = "region")
palatXregion <- getPalat (df, labels = "region", hex = "HEX.K")
print(palatXregion)

# Without labels nor hex codes
myColors <- data.frame(colors = c("gray36", "gray56"))
getPalat(myColors, hex = "colors")

}
