% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_if_needed.R
\name{install_if_needed}
\alias{install_if_needed}
\title{Install a package only if it hasn't before. Else/then load it}
\usage{
install_if_needed(pkg)
}
\arguments{
\item{pkg}{the name of the package in quotation marks}
}
\value{
NULL Package loaded
}
\description{
Install a package only if it hasn't before. Else/then load it
}
\examples{
#' install_if_needed("dplyr")

# List of packages you want to ensure are installed
packages <- c("devtools", "BiocManager", "dartRverse", "leaflet")
#' # Install and load each package
for (pkg in packages) {
  install_if_needed(pkg)
}

}
