% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plat_map.R
\name{plat_map}
\alias{plat_map}
\title{Map points given the latitude, longitude and color}
\usage{
plat_map(df, lat = "lat", lon = "lon", hex = "hex_color", size = 3, ...)
}
\arguments{
\item{df}{data frame with latitude, longitude, and color as values.}

\item{lat}{a vector, or column name (df$colname) for latitude values.}

\item{lon}{a vector, or column name (df$colname) for latitude values.}

\item{hex}{list of colors for each pair of coordinates}

\item{size}{default 3. Size of the points on the map}

\item{...}{Other arguments passed to geom_point()}
}
\value{
a map with the points generated from the coordinates and the color determined in the arguments
}
\description{
Map points given the latitude, longitude and color
}
\examples{

# Generate colors from coordinates
latitude = c(4.611, 6.251, 3.437, 10.391, 10.963)
longitude = c(-74.083, -75.563, -76.522, -75.514, -74.796)
df <- data.frame(latitude, longitude)
RGB_df <- getRGB(df, lat = "latitude", lon = "longitude")
HEX_df <- getHEX(df = RGB_df)
plat_map(df = HEX_df, lat = "latitude", lon = "longitude", alpha = 0.5)

colors_vector <- c("darkgoldenrod1", "gold", "dodgerblue3", "firebrick3", "ivory")
lat = c(4.611, 6.251, 3.437, 10.391, 10.963)
lon = c(-74.083, -75.563, -76.522, -75.514, -74.796)
df <- data.frame(lat, lon, colors_vector)
plat_map(df = df, hex = "colors_vector")

}
