% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.R
\name{getColors}
\alias{getColors}
\title{Add the R, G, B, and HEX columns to the input data frame. Optionally, map the results.}
\usage{
getColors(df, lat = "lat", lon = "lon", dir = "EW", map = FALSE, ...)
}
\arguments{
\item{df}{A data frame with latitude and longitude columns.}

\item{lat}{Column name that contains the latitude values. Default "lat".}

\item{lon}{Column name that contains the longitude values. Default "lon".}

\item{dir}{indicates the direction of the main axis, East - West ("EW") or
North - South ("NS"). Default "EW". This axis determine the Green - Red gradient from North or East, respectively.}

\item{map}{Logical. Whether a map is to be plotted. Default FALSE.}

\item{...}{Other arguments passed to geom_point() via platMap()}
}
\value{
A data frame with the original columns and values with the columns:
- Red, Green and Blue intensity values in decimal numbers (e.g., 100.2645).
- RGB color in decimal format (e.g., 155, 100, 160).
- RGB color in hexadecimal format (e.g., #9A649F).
if map = TRUE, returns a list with table and ggplot object.
}
\description{
Add the R, G, B, and HEX columns to the input data frame. Optionally, map the results.
}
\examples{
# Sample data frame
latitude <- c(4.611, 6.251, 3.437, 10.391, 10.963)
longitude <- c(-74.083, -75.563, -76.522, -75.514, -74.796)
dataframe <- data.frame(latitude, longitude)

# Call the function without plotting the map
getColors(df = dataframe, lat = "latitude", lon = "longitude", map = FALSE)

# Call the function and plot the map
getColors(df = dataframe, lat = "latitude", lon = "longitude")

}
