% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRGB.R
\name{getRGB}
\alias{getRGB}
\title{Calculate the red (R), green (G), and blue (B) color code from two or more values}
\usage{
getRGB(df, lat = "lat", lon = "lon", bind = TRUE)
}
\arguments{
\item{df}{data frame with latitude and longitude values. Must have at least two rows}

\item{lat}{the column name of df that contains the latitude values.}

\item{lon}{the column name of df that contains the longitude values.}

\item{bind}{logical. Whether the output data frame should be merged with the input data frame. Default TRUE}
}
\value{
A data frame containing RGB values calculated from the input data.
The output includes the following columns:
- R: Red channel values.
- G: Green channel values.
- B: Blue channel values.
Each row corresponds to a coordinate of latitude and longitude provided.
Each value represents an intensity level for the respective color channel.
}
\description{
Calculate the red (R), green (G), and blue (B) color code from two or more values
}
\examples{
# Using a data frame
latitude = c(4.611, 6.251, 3.437, 10.391, 10.963)
longitude = c(-74.083, -75.563, -76.522, -75.514, -74.796)
dataframe <- data.frame(latitude, longitude)
RGB_df <- getRGB(df = dataframe, lat = "latitude", lon = "longitude")
}
