% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRGB.R
\name{getRGB}
\alias{getRGB}
\title{Calculate the red (R), green (G), and blue (B) color code from two or more values}
\usage{
getRGB(lat, lon)
}
\arguments{
\item{lat}{a vector, or column name (df$colname) for latitude values. Must have at least two values}

\item{lon}{a vector, or column name (df$colname) for latitude values. Must have at least two values}
}
\value{
A numeric vector or matrix containing RGB values calculated from the input data.
The output includes the following components:
- R1, R2, R3, R4, R5: Red channel values.
- G1, G2, G3, G4, G5: Green channel values.
- B1, B2, B3, B4, B5: Blue channel values.
Each value represents an intensity level for the respective color channel.
}
\description{
Calculate the red (R), green (G), and blue (B) color code from two or more values
}
