% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplatPies.R
\name{ggplatPies}
\alias{ggplatPies}
\title{# Map pie charts given the latitude, longitude, color, proportions and groups}
\usage{
ggplatPies(
  df,
  lat = "lat",
  lon = "lon",
  k,
  hex = "HEX.K",
  mapid = "region",
  radius = 0.2,
  land = "gray",
  coast = "white",
  expand = 0.05,
  pie_border_col = "black",
  pie_line_size = 0.1,
  pie_alpha = 1,
  legend = "none"
)
}
\arguments{
\item{df}{data frame with latitude, longitude, and color as values.}

\item{lat}{the column name of df that contains the latitude values. Default "lat".}

\item{lon}{the column name of df that contains the longitude values. Default "lon".}

\item{k}{column name with the names of the clusters (e.g., group, region).
# The factors or elements in k must be equal as the colnames in the df that contains the proportions.}

\item{hex}{Vector or column name that contains the color relative to the cluster. Default "HEX.K".}

\item{mapid}{map used by ggplot2::geom_map. Default "region". Options are "subregion", "id", "state".}

\item{radius}{radius of the pie charts. Default 0.2.}

\item{land}{color of the land on the map. Default "gray".}

\item{coast}{color of the coast line (or country line). Default "white".}

\item{expand}{factor to expand the area of the map. Bigger pies require a bigger map area range.}

\item{pie_border_col}{color of the line that draws the pie chart. Default "black".}

\item{pie_line_size}{thickness of the line that draws the pie chart. Default 0.1.}

\item{pie_alpha}{transparency level of the pie chart}

\item{legend}{position of the legend. Default "none".}
}
\value{
a ggplot object with pie charts showing the proportion of each category with the colors given the location.
}
\description{
# Map pie charts given the latitude, longitude, color, proportions and groups
}
\examples{

# Sample data frame without the colors
df1 <- data.frame(
  lati = c(4.611, 6.251, 3.437, 10.391, 10.963, 7.984722),
  lon = c(-74.083, -75.563, -76.522, -75.514, -74.796, -75.198056),
  # [6] doesn't entirely belong to "Andes" or "Coast" region
  region = c("Andes", "Andes", "Andes", "Coast", "Coast", NA),
  # the following are the proportions.
  # Please not that the column names match the names of the region (cluster)
  Andes = c(1, 1, 1, 0, 0, 0.5),
  Coast = c(0, 0, 0, 1, 1, 0.5)
)
# Get the colors per sample and bind them to df
df1 <- getColors(df1, map = FALSE, lat = "lati")

# Get the average color per region
df1 <- colorCluster(df1, k = "region")
# the resultant df contains the necessary columns for ggplatPies.

# Plot the pie charts on a map
plot <- ggplatPies(df = df1, lat = "lati", k = "region", radius = 0.4,
                   expand = 0.15, pie_alpha = 1, legend = "right")
}
