% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_if_needed.R
\name{install_if_needed}
\alias{install_if_needed}
\title{Install packages if they have not been isntalled yet}
\usage{
install_if_needed(pkg)
}
\arguments{
\item{pkg}{The name of the package}
}
\value{
package installed and/or loaded.
}
\description{
Install packages if they have not been isntalled yet
}
\examples{
install_if_needed("dplyr")
# List of packages you want to ensure are installed

packages <- c("devtools", "BiocManager", "dartRverse", "leaflet")
#' # Install and load each package
for (pkg in packages) {
  install_if_needed(pkg)
}
}
