% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platMap.R
\name{platMap}
\alias{platMap}
\title{Map points given the latitude, longitude and color}
\usage{
platMap(df, lat = "lat", lon = "lon", hex = "HEX", size = 3, ...)
}
\arguments{
\item{df}{data frame with latitude, longitude, and color as values.}

\item{lat}{the column name of df that contains the latitude values. Default "lat".}

\item{lon}{the column name of df that contains the longitude values. Default "lon".}

\item{hex}{the column name of df that contains the color codes (or names). Default "HEX".}

\item{size}{default 3. Size of the points on the map}

\item{...}{Other arguments passed to geom_point()}
}
\value{
a map with the points generated from the coordinates and the color determined in the arguments
}
\description{
Map points given the latitude, longitude and color
}
\examples{

# Generate colors from coordinates
latitude = c(4.611, 6.251, 3.437, 10.391, 10.963)
longitude = c(-74.083, -75.563, -76.522, -75.514, -74.796)
df <- data.frame(latitude, longitude)
RGB_df <- getRGB(df, lat = "latitude", lon = "longitude")
HEX_df <- getHEX(df = RGB_df)
platMap(df = HEX_df, lat = "latitude", lon = "longitude", alpha = 0.5)

colors_vector <- c("#00FFD7", "gold", "dodgerblue3", "firebrick3", "ivory")
lat <- c(4.611, 6.251, 3.437, 10.391, 10.963)
lon <- c(-74.083, -75.563, -76.522, -75.514, -74.796)
df <- data.frame(lat, lon, colors_vector)
platMap(df = df, hex = "colors_vector")

}
