% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointCluster.R
\name{pointCluster}
\alias{pointCluster}
\title{Average RGB band color intensity within members of the same cluster}
\usage{
pointCluster(
  df,
  k,
  r = "Red",
  g = "Green",
  b = "Blue",
  lati = "lat",
  long = "lon",
  size = 2,
  label = TRUE
)
}
\arguments{
\item{df}{a data frame with numeric Red, Green and Blue values (three columns)
and a column that aggregates the points (e.g., Cluster)}

\item{k}{Column name with the information to aggregate/generalize the data
(e.g., cluster, population, group, region, ...). Default "Cluster".}

\item{r}{red channel column name. From 0 to 255. Default "Red"}

\item{g}{green channel column name. From 0 to 255. Default "Green"}

\item{b}{blue channel column name. From 0 to 255. Default "Blue"}

\item{lati}{the column name of df that contains the latitude values. Default "lat".}

\item{long}{the column name of df that contains the longitude values. Default "lon".}

\item{size}{size of the points to be plotted, passed through platMap()}

\item{label}{Whether the labels for k are to be printed. Default TRUE}
}
\value{
data frame with as many unique values as there are
in the cluster column.
The data frame contains six columns as follows:
- Cluster: the original name of the column is changed to Cluster,
- Red, Green and Blue: intensity of each color band respective to each
cluster.
- RGB: RGB color in decimal format respective to each cluster.
- HEX: RGB color in hexadecimal format respective to each cluster.
- lat: mean latitude respective to to each cluster.
- lon: mean longitude respective to to each cluster.
}
\description{
This function is useful when which to generaliz your map and get only one point per cluster.
However, it calculates the average coordinate of the points (not the centroid). Therefore,
should only be applied to generalize points that are not too far away from each other, so
the earth curvature does not have a meaningful effect. You must have the Red, Green and Blue values
for each row so the color is also computed. I suggest running getColors() first.
}
\examples{
# Sample data frame with the color band values
df <- data.frame(
lat = c(4.611, 6.251, 3.437, 10.391, 10.963),
long = c(-74.083, -75.563, -76.522, -75.514, -74.796),
clusters = c("Andes", "Andes", "Andes", "Coast", "Coast"),
Red = c(255, 200, 100, 50, 0),
Green = c(100, 150, 200, 255, 50),
Blue = c(50, 75, 100, 125, 150)
)

# Obtain a RGB colors per cluster only
colors <- pointCluster(df = df, k = "clusters", long = "long")
print(colors) # Only resultant values. One column per cluster.


# Workflow from latitude and longitude values
df <- data.frame(
lat = c(4.611, 6.251, 3.437, 10.391, 10.963),
long = c(-74.083, -75.563, -76.522, -75.514, -74.796),
clusters = c("Andes", "Andes", "Andes", "Coast", "Coast"))

# Get the RGB bands. This works also with getRGB(..., bind = TRUE)
df1 <- getColors(df = df, lon = "long", map = FALSE)
df1 <- pointCluster(df = df1, k = "clusters", long = "long", label = "clusters")


}
